Move getSignature to auth
