+++ app/auth/getSignature.ts
